+++ CHANGELOG.md
@@ -11,12 +11,14 @@ The format is based on the [KeepAChangeLog] project.
 - [#503] Fix error on UserInfo endpoint for removed clients
 - [#508] JWT now uses verify keys for JWT verification
 - [#502] IntrospectionEndpoint now returns False if it encounters any error as per specs
+- [#481] Loading AuthnEvent from session
 
 ### Removed
 - [#494] Methods and functions deprecated in previous releases have been removed
 
 ### Changed
 - [#507] Altered structure of client_db. It no longer stores mapping of ``registration_access_token`` to ``client_id``
+- [#481] AuthnEvent in session is now represented as JSON
 
 ### Added
 - [#496] Ability to specify additional supported scopes for oic.Provider
@@ -27,6 +29,7 @@ The format is based on the [KeepAChangeLog] project.
 [#508]: https://github.com/OpenIDC/pyoidc/issues/508
 [#507]: https://github.com/OpenIDC/pyoidc/issues/507
 [#502]: https://github.com/OpenIDC/pyoidc/issues/502
+[#481]: https://github.com/OpenIDC/pyoidc/issues/481
 
 ## 0.13.1 [2018-04-06]